+++ src/bloggy_frontend/src/pages/Posts.jsx
@@ -1,25 +1,12 @@
 import React from 'react'
-import { Container, Row, Col, Card, Button, Spinner } from 'react-bootstrap'
-import { useQuery } from '@tanstack/react-query'
-import { bloggy_backend } from '../../../declarations/bloggy_backend'
+import { Container, Row, Col, Spinner, Button } from 'react-bootstrap'
 import { Link } from 'react-router-dom'
 import { useAuth } from '../context/AuthContext'
-import { QUERY_KEYS } from '../constants/queryKeys'
-
-const fetchPosts = async () => {
-  const posts = bloggy_backend.viewPosts()
-  return posts
-}
+import PostCard from '../components/PostCard'
+import { useFetchPosts } from '../hooks/useFetchPosts'
 
 const Posts = () => {
-  const {
-    data: posts,
-    error,
-    isLoading,
-  } = useQuery({
-    queryKey: [QUERY_KEYS.POSTS],
-    queryFn: fetchPosts,
-  })
+  const { data: posts, error, isLoading } = useFetchPosts()
   const { logOut } = useAuth()
 
   return (
@@ -38,9 +25,7 @@ const Posts = () => {
       </div>
       {isLoading ? (
         <div className="text-center mt-5">
-          <Spinner animation="border" role="status">
-            <span className="sr-only">Loading...</span>
-          </Spinner>
+          <Spinner animation="border" role="status"></Spinner>
         </div>
       ) : error ? (
         <div className="text-danger text-center mt-5">{error.message}</div>
@@ -48,20 +33,7 @@ const Posts = () => {
         <Row xs={1} md={2} lg={3} className="g-4">
           {posts.map((post, index) => (
             <Col key={index}>
-              <Card>
-                <Card.Body>
-                  <Card.Title>{post.title}</Card.Title>
-                  <Card.Text>{post.description}</Card.Text>
-                  <div className="d-flex justify-content-between align-items-center">
-                    <div>
-                      <span role="img" aria-label="upvote">
-                        👍
-                      </span>
-                      {post.upvotes}
-                    </div>
-                  </div>
-                </Card.Body>
-              </Card>
+              <PostCard post={post} />
             </Col>
           ))}
         </Row>
